+++ packages/graph-explorer/tailwind.config.ts
@@ -107,6 +107,10 @@ export default {
       base: "1rem",
       lg: "1.125rem",
       xl: "1.25rem",
+      "2xl": "1.5rem",
+      "3xl": "1.75rem",
+      "4xl": "2rem",
+      "5xl": "2.5rem",
     },
     fontWeight: {
       light: "400",