+++ Changelog.md
@@ -4,6 +4,8 @@
 
 - **Improved** performance of styling sidebar panels when many node & edge types
   exist ([#542](https://github.com/aws/graph-explorer/pull/542))
+- **Added** global error page if the React app crashes
+  ([#547](https://github.com/aws/graph-explorer/pull/547))
 - **Transition** to Tailwind instead of EmotionCSS for styles, which should make
   updating the UI much simpler
   ([#543](https://github.com/aws/graph-explorer/pull/543))